refactor: Redesign template system for scalability
- Extract common template logic to BaseTemplate
- Standardize component structure and initialization
- Implement namespace-based component discovery
- Simplify creation of new templates with minimal code
