+++ src/JATSParser/PDF/TemplateStrategy.php
@@ -1,18 +1,11 @@
 <?php namespace JATSParser\PDF;
 
-use JATSParser\PDF\PDFConfig\Configuration;
-use JATSParser\PDF\TCPDFDocument;
-use JATSParser\PDF\TemplateOne;
-
-require_once __DIR__ . '/Templates/TemplateOne.php';
-require_once __DIR__ . '/PDFConfig/Configuration.php';
-
 class TemplateStrategy {
 
     private $pdfDocument;
 
     function __construct(string $templateName, $config) {
-        $namespace = "JATSParser\\PDF\\Templates\\$templateName";
+        $namespace = "JATSParser\\PDF\\Templates\\$templateName\\$templateName";
         $this->pdfDocument = new $namespace($config);
     }
