+++ src/JATSParser/PDF/Templates/TemplateOne/TemplateOne.php
@@ -1,0 +1,11 @@
+<?php namespace JATSParser\PDF\Templates\TemplateOne;
+
+use JATSParser\PDF\Templates\BaseTemplate;
+
+use JATSParser\PDF\Templates\TemplateOne\Components\TemplateBody;
+use JATSParser\PDF\Templates\TemplateOne\Components\Header;
+use JATSParser\PDF\Templates\TemplateOne\Components\Footer;
+use JATSParser\PDF\Templates\TemplateOne\Components\Body;
+
+class TemplateOne extends BaseTemplate {
+}
